Update Mono_to_Hydrazine.cfg
Would it look better if the "// this does this..." parameters were simply put a couple tabs away from each of their parent? ATM I have them (when done in NotePad++, which I like to use for coding assistance and visualization) all aligned the same, HOWEVER, when viewed in really anything else it looks messy. So here in this edit I put them 2 tabs away each.

Should I keep it this way?