Update Stockalike_RLA.cfg
RLA Stockalike 13 changed part names and modified a few engines. I tried to correct them here.

The 88-M6 "Spinnaker" is supposed to be the low efficiency high thrust engine in the small rockomax family. As it was configured it seemed to be worse than the 48-7S (liquidEngineMini) in every way. Made it an "L" engine with a kerolox thrust of 90, putting it at the top end of the small engines, power wise.

I cannot seem to get the Engine Ignitions working on these :-(. Can you take a look?