Fix MM syntax on CT-10 Chelyabinsk plume
The notation "@CONFIG[Kerolox,Methalox]" doesn't work to apply the same patch
to CONFIG nodes with both of those names, like it may have in the past. In
current ModuleManager, it tries to patch a CONFIG node that has both
"name=Kerolox" and "tag=Methalox".

Since the CT-10 doesn't actually have a kerolox mode in the current
configuration, I deleted the reference to Kerosene+LqdOxygen from the plume
configuration.