Update Squad_NTR_modularEngines.cfg
ModuleHybridEngine should not be used with MultiModeEngine. It's also somewhat outdated. I can't see a need for it unless you want to be able to configure an engine for multiple propellants AND also have each config support afterburner mode. (I think that is possible but not quite sure and these configs don't attempt that anyway)

History: ModuleHybridEngine is an extension of ModuleEngineConfigs that provided multi mode support before MultiModeEngine module was added to stock. It's somewhat obsolete except for the aforementioned possibility of creating multiple propellant/afterburner combinations. Assuming that actually works...