Allow 2 ignitions for RE-I2 in Hydrolox mode
The [Saturn V S-IVB](https://en.wikipedia.org/wiki/Saturn_V#S-IVB_third_stage) third stage runs on Hydrolox and has 2 ignitions - one for orbital insertion and one for translunar injection. Updated the ignition count for Hydrolox to match this. If there's a way to toggle that value, instead of forcing it to always be 2, I'll gladly do that instead.